Move out require setup
