+++ lute/cli/CMakeLists.txt
@@ -28,6 +28,7 @@ target_sources(Lute.CLI.lib PRIVATE
     include/lute/fileutils.h
     include/lute/luauflags.h
     include/lute/reporter.h
+    include/lute/requiresetup.h
     include/lute/staticrequires.h
     include/lute/tc.h
     include/lute/uvstate.h
@@ -36,6 +37,7 @@ target_sources(Lute.CLI.lib PRIVATE
     src/compile.cpp
     src/fileutils.cpp
     src/luauflags.cpp
+    src/requiresetup.cpp
     src/staticrequires.cpp
     src/tc.cpp
     src/uvstate.cpp
