+++ lute/cli/src/climain.cpp
@@ -1,41 +1,27 @@
 #include "lute/climain.h"
 
-#include "lute/bundlevfs.h"
 #include "lute/clicommands.h"
-#include "lute/clivfs.h"
 #include "lute/compile.h"
-#include "lute/crypto.h"
-#include "lute/fs.h"
-#include "lute/io.h"
-#include "lute/luau.h"
 #include "lute/luauflags.h"
-#include "lute/net.h"
+#include "lute/modulepath.h"
 #include "lute/options.h"
 #include "lute/process.h"
 #include "lute/ref.h"
 #include "lute/reporter.h"
-#include "lute/require.h"
-#include "lute/requirevfs.h"
+#include "lute/requiresetup.h"
 #include "lute/runtime.h"
 #include "lute/staticrequires.h"
-#include "lute/system.h"
-#include "lute/task.h"
 #include "lute/tc.h"
-#include "lute/time.h"
 #include "lute/version.h"
-#include "lute/vm.h"
 
 #include "Luau/CodeGen.h"
 #include "Luau/Common.h"
 #include "Luau/Compiler.h"
 #include "Luau/DenseHash.h"
 #include "Luau/FileUtils.h"
-#include "Luau/Require.h"
 
 #include "lua.h"
 #include "lualib.h"
-
-#include "uv.h"
 
 #include <memory>
 
@@ -115,120 +101,6 @@ Compile Options:
 	--show-require-graph    Print the require dependency graph.
 	-h, --help              Display this usage message.
 )";
-
-void* createCliRequireContext(lua_State* L)
-{
-    void* ctx = lua_newuserdatadtor(
-        L,
-        sizeof(RequireCtx),
-        [](void* ptr)
-        {
-            std::destroy_at(static_cast<RequireCtx*>(ptr));
-        }
-    );
-
-    if (!ctx)
-        luaL_error(L, "unable to allocate RequireCtx");
-
-    ctx = new (ctx) RequireCtx{std::make_unique<RequireVfs>(CliVfs{})};
-
-    // Store RequireCtx in the registry to keep it alive for the lifetime of
-    // this lua_State. Memory address is used as a key to avoid collisions.
-    lua_pushlightuserdata(L, ctx);
-    lua_insert(L, -2);
-    lua_settable(L, LUA_REGISTRYINDEX);
-
-    return ctx;
-}
-
-static void luteopen_libs(lua_State* L)
-{
-    std::vector<std::pair<const char*, lua_CFunction>> libs = {{
-        {"@lute/crypto", luteopen_crypto},
-        {"@lute/fs", luteopen_fs},
-        {"@lute/luau", luteopen_luau},
-        {"@lute/net", luteopen_net},
-        {"@lute/process", luteopen_process},
-        {"@lute/task", luteopen_task},
-        {"@lute/vm", luteopen_vm},
-        {"@lute/system", luteopen_system},
-        {"@lute/time", luteopen_time},
-        {"@lute/io", luteopen_io},
-    }};
-
-    for (const auto& [name, func] : libs)
-    {
-        lua_pushcfunction(L, luarequire_registermodule, nullptr);
-        lua_pushstring(L, name);
-        func(L);
-        lua_call(L, 2, 0);
-    }
-}
-
-void* createBundleRequireContext(
-    lua_State* L,
-    Luau::DenseHashMap<std::string, std::string> luaurcFiles,
-    Luau::DenseHashMap<std::string, std::string> bundleMap
-)
-{
-    void* ctx = lua_newuserdatadtor(
-        L,
-        sizeof(RequireCtx),
-        [](void* ptr)
-        {
-            std::destroy_at(static_cast<RequireCtx*>(ptr));
-        }
-    );
-
-    if (!ctx)
-        luaL_error(L, "unable to allocate RequireCtx");
-    ctx = new (ctx) RequireCtx{std::make_unique<RequireVfs>(BundleVfs{std::move(luaurcFiles), std::move(bundleMap)})};
-
-    // Store RequireCtx in the registry to keep it alive for the lifetime of
-    // this lua_State. Memory address is used as a key to avoid collisions.
-    lua_pushlightuserdata(L, ctx);
-    lua_insert(L, -2);
-    lua_settable(L, LUA_REGISTRYINDEX);
-
-    return ctx;
-}
-
-lua_State* setupCliState(Runtime& runtime, std::function<void(lua_State*)> preSandboxInit)
-{
-    return setupState(
-        runtime,
-        [preSandboxInit = std::move(preSandboxInit)](lua_State* L)
-        {
-            luteopen_libs(L);
-
-            if (Luau::CodeGen::isSupported())
-                Luau::CodeGen::create(L);
-
-            luaopen_require(L, requireConfigInit, createCliRequireContext(L));
-            if (preSandboxInit)
-                preSandboxInit(L);
-        }
-    );
-}
-
-lua_State* setupBundleState(
-    Runtime& runtime,
-    Luau::DenseHashMap<std::string, std::string> luaurcFiles,
-    Luau::DenseHashMap<std::string, std::string> bundleMap
-)
-{
-    return setupState(
-        runtime,
-        [luaurcFiles = std::move(luaurcFiles), bundleMap = std::move(bundleMap)](lua_State* L)
-        {
-            luteopen_libs(L);
-            if (Luau::CodeGen::isSupported())
-                Luau::CodeGen::create(L);
-
-            luaopen_require(L, requireConfigInit, createBundleRequireContext(L, std::move(luaurcFiles), std::move(bundleMap)));
-        }
-    );
-}
 
 static bool setupArguments(lua_State* L, int argc, char** argv)
 {
