+++ tests/src/cliruntimefixture.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "lute/climain.h"
+
 #include "lute/runtime.h"
 
 #include "lutefixture.h"
